Create SpaceTravel.py
ChatGPT comment:
Your code looks good to me! You have defined a parent class called "Spacecraft" with attributes and methods, and two child classes called "Spaceship" and "Probe" that inherit from the parent class and have additional attributes and methods. You have also created instances of both child classes and tested all the methods to ensure they return the expected output.

One minor thing I noticed is that in the "travel" method of the "Spacecraft" class, you are multiplying the result by 60 to convert the travel time from minutes to seconds. It might be more clear to add a comment explaining this conversion, since it might not be immediately obvious why you are multiplying by 60.

Other than that, great job!